chunk range is not dependable, not sure why
if we dont get a chunk range, we have to be OK about that, there is some places
in the conf test suite its sent and some its not, i'll dig in to figure out when
the spec says to send it, but the way its modeled now, we just accept empty uploads (unless this is another failing test I havent gotten to yet).

When we verify, we just assume :ok in this case. I'll do some research on the spec so when its required is more clearn, right now it looks like a hack and it'll definitely
permit someone to make a million empty files and it be 'valid' (and expensive) over something like s3.